Initial Push

+++ repo_aktualisieren.sh
@@ -27,7 +27,10 @@ cd $REPO
 echo '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' >$REPO/addons.xml
 echo '<addons>' >> $REPO/addons.xml
 for name in `find . -maxdepth 1 -type d |grep -v \.git|grep -v addons|egrep -v "^\.$"|cut -d \/ -f 2 `; do
-   cp -R -p ~/kodi/addons/$name/* $name
+   echo "Adding $name"
+   if [ -f "~/kodi/addons/$name/addon.xml" ]; then
+     cp -R -p ~/kodi/addons/$name/* $name
+   fi
    find $name -iname "*.pyo" -exec rm {} \;
    find $name -iname ".DS_Store" -exec rm {} \;
    VERSION=`cat $name/addon.xml|grep \<addon|grep $name |tr 'A-Z' 'a-z'|sed 's/.*version="\([^"]*\)"*.*/\1/g'`
